% wll-12-10-2006: First draft
% dle-drp-15-09-2007: Major changes

\name{AnnotationDB}
\docType{data}
\alias{AnnotationDB}

\title{Annotation database}
\usage{
data(AnnotationDB)
}

\description{
 Annotation database
}

\value{
Data frame:
 \item{GName}{GName}
 \item{Class}{Chemical class}
 \item{MF}{Molecular formula}
 \item{MW}{Exact monoisotopic mass}
 \item{HMDB}{HMDB}
 \item{KEGG}{KEGG}
 \item{PC}{PC}"  
}


\author{
David Enot \email{david.enot@gustaveroussy.fr}
}


\examples{
str(AnnotationDB)
}    
\keyword{datasets}
