\name{Undocumented functions}
\alias{.GRgetArea}
\alias{.GRmsExtrema}
\alias{.GRdoksmooth}
\alias{.GRipeaks}
\alias{.GRmsPeakSimple}
\alias{.GRconvEIC}
\alias{.GRparseSampEICCl}
\alias{.GRgatherMultiEICCl}
\alias{.GRdoOneCef}
\alias{.GRgetcomptime}
\title{Undocumented functions}
\description{

\code{.GRgetArea} integrate area

\code{.GRmsExtrema} find extrema in chromatogram

\code{.GRipeaks}

\code{.GRmsPeakSimple} basic peak finding

\code{.GRconvEIC} align/smooth an EIC across samples

\code{.GRdoksmooth} in .GRconvEIC

\code{.GRparseSampEICCl} for parallel parseSampEIC

\code{.GRgatherMultiEICCl} for parallel gatherMultiEIC

\code{.GRdoOneCef} parse one .cef file

\code{.GRgetcomptime} extract completion time in mzdata.xml file - few sec per file

These functions are currently undocumented.
}
\keyword{internal}% at least one, from doc/KEYWORDS

