% wll-12-10-2006: First draft
% dle-drp-15-09-2007: Major changes

\name{RatiosGR}
\docType{data}
\alias{RatiosGR}

\title{Annotation database of ratios}
\usage{
data(RatiosGR)
}

\description{
 Annotation database
}

\value{
List with name of the ratio and numerator / denominator
}


\author{
David Enot \email{david.enot@gustaveroussy.fr}
}


\examples{
str(RatiosGR)
}    
\keyword{datasets}
