\name{Undocumented functions}
\alias{.GRgetArea}
\alias{.GRmsExtrema}
\alias{.GRdoksmooth}
\alias{.GRipeaks}
\alias{.GRmsPeakSimple}
\alias{.GRconvEIC}
\alias{.GRparseSampEICCl}
\alias{.GRgatherMultiEICCl}
\alias{.GRdoOneCef}
\alias{.GRgetcomptime}
\title{Undocumented functions}
\description{

\code{.GRgetArea}

\code{.GRmsExtrema}

\code{.GRdoksmooth}

\code{.GRipeaks}

\code{.GRmsPeakSimple}

\code{.GRconvEIC}

\code{.GRparseSampEICCl}

\code{.GRgatherMultiEICCl}

\code{.GRdoOneCef}

\code{.GRgetcomptime}

These functions are currently undocumented.
}
\keyword{internal}% at least one, from doc/KEYWORDS

