\name{Undocumented functions}
\alias{.GRgetArea}
\alias{.GRmsExtrema}
\alias{.GRdoksmooth}
\alias{.GRipeaks}
\alias{.GRmsPeakSimple}
\alias{.GRconvEIC}
\title{Undocumented functions}
\description{

\code{.GRgetArea}

\code{.GRmsExtrema}

\code{.GRdoksmooth}

\code{.GRipeaks}

\code{.GRmsPeakSimple}

\code{.GRconvEIC}

These functions are currently undocumented.
}
\keyword{internal}% at least one, from doc/KEYWORDS

